PROD-1566 - terms, privacy policy, and order contract modals - style
