+++ src-ts/lib/form/form-inputs/form-input-row/FormInputRow.tsx
@@ -46,7 +46,7 @@ const FormInputRow: (props: FormInputRowProps) => JSX.Element = (props: FormInpu
         )
 
     return (
-        <div>
+        <div className={styles['row-wrap']}>
             {title}
             {inputRow}
         </div>
