+++ src-ts/lib/modals/index.ts
@@ -1,2 +1,5 @@
 export * from './base-modal'
 export * from './confirm'
+export * from './order-contract-modal'
+export * from './privacy-policy-modal'
+export * from './terms-modal'
