+++ src-ts/lib/modals/order-contract-modal/OrderContractModal.module.scss
@@ -1,0 +1,24 @@
+@import '../../styles/palette';
+
+.container {
+  ol {
+    padding: 0;
+  }
+
+  h4 {
+    margin-top: 20px;
+  }
+  
+  p {
+    margin-bottom: 20px;
+
+    &.sm {
+      font-size: 14px;
+    }
+  }
+}
+
+.topCoderLink {
+  text-decoration: underline;
+  color: $link-blue-light;
+}
