+++ src-ts/lib/modals/order-contract-modal/index.ts
@@ -1,0 +1,1 @@
+export { default as OrderContractModal } from './OrderContractModal'
