+++ src-ts/lib/modals/privacy-policy-modal/index.ts
@@ -1,0 +1,1 @@
+export { default as PrivacyPolicyModal } from './PrivacyPolicyModal'
