+++ src-ts/lib/modals/terms-modal/index.ts
@@ -1,0 +1,1 @@
+export { default as TermsModal } from './TermsModal'
