+++ src-ts/lib/page-footer/PageFooter.tsx
@@ -1,7 +1,8 @@
 
-import { FC } from 'react'
+import { Dispatch, FC, SetStateAction, useState } from 'react'
 
 import { FooterSocialConfig } from '../../config'
+import { OrderContractModal, PrivacyPolicyModal, TermsModal } from '../modals'
 import { SocialLink, SocialLinkIcons } from '../social-links'
 
 import styles from './PageFooter.module.scss'
@@ -13,8 +14,16 @@ export interface PageFooterProps {
 
 const PageFooter: FC<PageFooterProps> = ({
 }: PageFooterProps) => {
+    const [isOrderContractModalOpen, setIsOrderContractModalOpen]: [boolean, Dispatch<SetStateAction<boolean>>] = useState<boolean>(false)
+    const [isPrivacyModalOpen, setIsPrivacyModalOpen]: [boolean, Dispatch<SetStateAction<boolean>>] = useState<boolean>(false)
+    const [isTermsModalOpen, setIsTermsModalOpen]: [boolean, Dispatch<SetStateAction<boolean>>] = useState<boolean>(false)
+
     return (
         <div className={styles['footer-wrap']}>
+            <OrderContractModal isOpen={isOrderContractModalOpen} onClose={() => setIsOrderContractModalOpen(false)} />
+            <PrivacyPolicyModal isOpen={isPrivacyModalOpen} onClose={() => setIsPrivacyModalOpen(false)} />
+            <TermsModal isOpen={isTermsModalOpen} onClose={() => setIsTermsModalOpen(false)} />
+
             <div className={styles['footer-inner']}>
                 <div className={styles.utils}>
                     <div>
@@ -23,8 +32,8 @@ const PageFooter: FC<PageFooterProps> = ({
                         <a href='#'>See a Bug?</a>
                     </div>
                     <div>
-                        <a href='#'>Terms</a>
-                        <a href='#'>Privacy Policy</a>
+                        <a href='#' onClick={(e) => {e.preventDefault(); setIsTermsModalOpen(true)}}>Terms</a>
+                        <a href='#' onClick={(e) => {e.preventDefault(); setIsPrivacyModalOpen(true)}}>Privacy Policy</a>
                     </div>
                 </div>
                 <div className={styles.social}>
