--- src/components/Modal/OrderContract/styles.module.scss
@@ -1,14 +1,0 @@
-@import "styles/include";
-
-.container {
-  font: 16px;
-}
-
-li {
-  padding-left: 20px;
-}
-
-.topCoderLink {
-  text-decoration: underline;
-  color: $link-blue!important;
-}
